fix: enhance error handling in authentication route by adding catch for API call and updating success condition

+++ src/routes/_authenticated/route.tsx
@@ -5,8 +5,13 @@ import { toast } from "sonner";
 export const Route = createFileRoute("/_authenticated")({
   component: RouteComponent,
   loader: async ({ location }) => {
-    const { data } = await api.post(`/api/users/validate`);
-    if (!data.success) {
+    const { data, status } = await api.post(`/api/users/validate`).catch(() => {
+      return {
+        data: { success: false, message: "Invalid token" },
+        status: 401,
+      };
+    });
+    if (!data.success && status !== 200) {
       toast.error(data.message);
       localStorage.clear();
       throw redirect({ to: "/", search: { redirect: location.href } });
